requirements.txt
pandas==1.2.3
python-dotenv==0.20.0
requests==2.27.1
selenium==4.1.5
undetected_chromedriver==3.0.6
